requirements.txt
# Core ML
torch==2.2.2
torchvision==0.17.2
timm==0.9.16

# Experiment tracking
wandb==0.16.6

# API serving
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Gradio UI
gradio==4.31.4

# Monitoring
evidently==0.4.22
scikit-image==0.23.2
Pillow==10.3.0
opencv-python-headless==4.9.0.80

# Data / utils
numpy==1.26.4
pandas==2.2.2
scipy==1.13.0

# Testing
pytest==8.2.0
httpx==0.27.0

# Misc
python-dotenv==1.0.1
tqdm==4.66.4
